Move the NullDuration to internal/database package;
