+++ internal/database/types_test.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package domain
+package database
 
 import (
 	"time"
@@ -17,24 +17,24 @@ type typesSuite struct {
 
 var _ = gc.Suite(&typesSuite{})
 
-func (s *typesSuite) TestNullableDuration(c *gc.C) {
-	nd := NullableDuration{Duration: 10 * time.Second, Valid: true}
+func (s *typesSuite) TestNullDuration(c *gc.C) {
+	nd := NullDuration{Duration: 10 * time.Second, Valid: true}
 	v, err := nd.Value()
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(v, gc.Equals, int64(10*time.Second))
 
-	nd = NullableDuration{Valid: true}
+	nd = NullDuration{Valid: true}
 	v, err = nd.Value()
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(v, gc.Equals, int64(0))
 
-	nd = NullableDuration{Valid: false}
+	nd = NullDuration{Valid: false}
 	v, err = nd.Value()
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(v, gc.IsNil)
 
 	err = nd.Scan("10s")
-	c.Assert(err, gc.ErrorMatches, `cannot scan type string into NullableDuration`)
+	c.Assert(err, gc.ErrorMatches, `cannot scan type string into NullDuration`)
 	c.Assert(nd.Valid, jc.IsFalse)
 
 	err = nd.Scan(int64(20 * time.Second))
